\name{RPPASet-class}
\alias{RPPASet-class}
\alias{RPPASet}
\alias{summary,RPPASet-method}
\alias{image,RPPASet-method}
\docType{class}
\title{The RPPASet Class}
\description{
  The RPPASet class fits supercurves to an entire directory of
  reverse-phase protein array experiments.
}
\usage{
RPPASet(path, designparams, fitparams, blanks=NULL)
\S4method{summary}{RPPASet}(object, path, prefix="supercurve", graphs=TRUE, tiffdir=TRUE, \dots)
}
\arguments{
  \item{path}{character string specifying the path from the current
    directory to the directory containing the files to be processed.}
  \item{designparams}{An RPPADesignParams object desctribingte desing
    common to all files.}
  \item{fitparams}{A RPPAFitParams object containing the parameters used
  to fit the super curve model.}
  \item{blanks}{Logical vector indicating the position of blankd
    dilution series. DEPRECATED.}
  \item{object}{An RPPASet object.}
  \item{prefix}{Character string used as a prefix on files generated by
    the summary method.}
  \item{graphs}{Logical: should the summary produce standard graphs.}
  \item{tiffdir}{Path to the directory containing the original TIFF files.}
  \item{\dots}{extra arguments for generic or plotting routines}
}
\details{
  Quantify all the slides in a directory using RPPASet. This returns an object
  containing slide data and fits for each slide.    Typically this is
  followed by a call to \code{summary} or \code{write.summary} to write
  the resulting quantifications and diagnostic plots to a directory.
}
\section{Objects from the Class}{
  Although objects of the class can (in theory) be created by a direct
  call to \link{new},  the only realistic  method is to use the
  \code{RPPASet}  generator function. 
}
\section{Slots}{
  \describe{
    \item{\code{call}:}{Record of the function call used during contruction}
    \item{\code{vresion}:}{Character string containing the version of the
      package used to construct the object.}
    \item{\code{design}:}{The RPPADesign object, common to all the slides.}
    \item{\code{rppas}:}{An array of RPPA objects.}
    \item{\code{fitparams}:}{The RPPAFitParams object that was used to
      construct the model fits.}
    \item{\code{fits}:}{An array of fitted objects.}
  }
}
\section{Methods}{
  \describe{
    \item{summary(object, path, prefix, graphs, tiffdir, \dots)}{The
      summary method creates a record of the entire RPPASet, including
      fitted values, resuiduals, and images of the processed slides. See
      the documentation of \code{write.summary} for details.}
  }
}
\value{
  The \code{RPPASet} generator returns an object of the \code{RPPASet} class.

  The \code{summary} method returns a summary of the underlying data
  frame.
}
\references{KRC}
\author{Kevin R. Coombes <kcoombes@mdanderson.org>}
\seealso{
  \code{\link{RPPAFit}},
  \code{\link{RPPADesign}}
}
\examples{
  \dontrun{
designparams <- RPPADesignParams(grouping="blockSample",
                                 center=FALSE,
                                 controls=list("control",
                                               "pos con",
                                               "neg con"))
fitparams <- RPPAFitParams(measure="Mean.Net",
                           method="nlrob",
                           model="cobs",
                           ignoreNegative=FALSE,
                           warnLevel=-1,
                           verbose=FALSE)

parentdir <- "Y:/MyData"
txtdir <- file.path(parentdir, "txt")     # quantification files
imgdir <- file.path(parentdir, "tif")     # and corresponding image files
outdir <- file.path(parentdir, "results") # output files

fitset <- RPPAFitDir(txtdir, designparams, fitparams)
write.summary(fitset,
              namebase="supercurve",
              path=outdir,
              normalize="median",
              graphs=TRUE,
              tiffdir=imgdir)
  }
}
\keyword{models}
\keyword{regression}
\keyword{nonlinear}
\keyword{nonparametric}
\keyword{robust}


  
