% $Id$

\name{SuperCurve-package}
\alias{SuperCurve-package}
\docType{package}
\title{Reverse phase protein lysate array analysis}
\description{
  A package for analyzing reverse phase protein lysate arrays (RPPA).
}
\details{
  \tabular{ll}{
    Package: \tab SuperCurve\cr
    Type:    \tab Package\cr
    Version: \tab 1.4.5\cr
    Date:    \tab 2012-06-01\cr
    License: \tab ARtistic-2.0\cr
  }

  For a complete list of functions, use \code{library(help="SuperCurve")}.
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="SuperCurve"))}.
}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

