\name{RPPASet-class}
\alias{RPPASet-class}
\alias{RPPASet}
\alias{summary,RPPASet-method}
\docType{class}
\title{Class "RPPASet"}
\description{
  The RPPASet class fits supercurves to an entire directory of
  reverse-phase protein array experiments.
}
\usage{
RPPASet(path, designparams, fitparams, software="microvigene")
\S4method{summary}{RPPASet}(object, path, prefix="supercurve", graphs=TRUE, tiffdir=TRUE, \dots)
}
\arguments{
  \item{path}{character string specifying the path from the current
    directory to the directory containing the files to be processed}
  \item{designparams}{object of class \code{RPPADesignParams} describing
    features common to all quantification files}
  \item{fitparams}{object of class \code{RPPAFitParams} containing
    parameters used to fit the supercurve model}
  \item{software}{character string specifying the software used to generate
    the quantification file (see section \sQuote{Details} of
    \code{\link{RPPA}})}
  \item{object}{object of class \code{RPPASet}}
  \item{prefix}{character string used as a prefix on files generated by
    the summary method.}
  \item{graphs}{logical scalar specifying if standard graphs should be produced}
  \item{tiffdir}{character string specifying path to the directory containing
    the original TIFF files}
  \item{\dots}{extra arguments for generic or plotting routines}
}
\details{
  Quantify all the slides in a directory using RPPASet. This returns an object
  containing slide data and fits for each slide. Typically this is followed by
  a call to \code{summary} or \code{write.summary} to write the resulting
  quantifications and diagnostic plots to a directory.
}
\section{Objects from the Class}{
  Although objects of the class can (in theory) be created by a direct call
  to \link[methods]{new}, the only realistic method is to use the
  \code{RPPASet} generator function.
}
\section{Slots}{
  \describe{
    \item{\code{call}:}{object of class \code{call} specifying the function
      call that was used during construction}
    \item{\code{version}:}{character string containing the version of this
      package used to construct the object}
    \item{\code{design}:}{object of class \code{RPPADesign}, common to all
      the slides}
    \item{\code{rppas}:}{array of objects of class \code{RPPA}}
    \item{\code{fitparams}:}{object of class \code{RPPAFitParams} that was
      used to construct the model fits}
    \item{\code{fits}:}{array of fitted objects of class \code{RPPAFit}}
  }
}
\section{Methods}{
  \describe{
    \item{summary(object, path, prefix, graphs, tiffdir, \dots)}{Creates
      a record of the entire RPPASet, including fitted values, residuals,
      and images of the processed slides. See the documentation of
      \code{write.summary} for details.}
  }
}
\value{
  The \code{RPPASet} generator returns an object of class \code{RPPASet}.

  The \code{summary} method returns a summary of the underlying data frame.
}
\references{KRC}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\seealso{
  \code{\link{RPPA}},
  \code{\link{RPPADesign}},
  \code{\link{RPPAFit}}
}
\examples{
  \dontrun{
designparams <- RPPADesignParams(grouping="blockSample",
                                 center=FALSE,
                                 controls=list("control",
                                               "pos con",
                                               "neg con"))
fitparams <- RPPAFitParams(measure="Mean.Net",
                           method="nlrob",
                           model="cobs",
                           ignoreNegative=FALSE,
                           warnLevel=-1,
                           verbose=FALSE)

parentdir <- file.path("C:", "MyData")
txtdir <- file.path(parentdir, "txt")     # quantification files
imgdir <- file.path(parentdir, "tif")     # and corresponding image files
outdir <- file.path(parentdir, "results") # output files

fitset <- RPPASet(txtdir, designparams, fitparams)
write.summary(fitset,
              path=outdir,
              graphs=TRUE,
              tiffdir=imgdir)
  }
}
\keyword{models}
\keyword{regression}
\keyword{nonlinear}
\keyword{nonparametric}
\keyword{robust}

