\name{qcprob-method}
\alias{qcprob}
\alias{qcprob-method}
\alias{qcprob,ANY-method}
\docType{methods}
\title{Method "qcprob"}
\description{
  Returns a numeric scalar representing the probability of a good slide.
}
\value{
  The \code{qcprob} method returns a numeric scalar.
}
\author{
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\keyword{methods}

