\name{elapsed-method}
\alias{elapsed}
\alias{elapsed-method}
\alias{elapsed,ANY-method}
\docType{methods}
\title{Method "elapsed"}
\description{
  Returns the elapsed time since an operation began. 
}
\usage{
elapsed(object,
        \dots)
}
\arguments{
  \item{object}{object of implementing class}
  \item{\dots}{extra arguments for generic routines}
}
\author{
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\keyword{methods}

