\name{paramString-method}
\alias{paramString}
\alias{paramString-method}
\alias{paramString,ANY-method}
\docType{methods}
\title{Method "paramString"}
\description{
  Returns a string representation of an object's slots.
}
\value{
  The \code{paramString} method returns a character vector, possibly
  empty but never \code{NULL}.
}
\section{Warning}{
  The \code{paramString} method should not be called by user except for
  informational purposes. The content and format of the returned string
  may vary between different versions of this package.
}
\author{
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\keyword{methods}
\keyword{internal}

