\name{SuperCurve-package}
\alias{SuperCurve-package}
\docType{package}
\title{Reverse phase protein lysate array analysis}
\description{
  A package for analyzing reverse phase protein lysate arrays (RPPA).
}
\details{
  \tabular{ll}{
    Package: \tab SuperCurve\cr
    Type: \tab Package\cr
    Version: \tab 1.4.3\cr
    Date: \tab 2012-03-16\cr
    License: \tab See LICENSE for license information\cr
  }

  For a complete list of functions, use \code{library(help="SuperCurve")}.
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="SuperCurve"))}.
}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\keyword{package}

