\name{SuperCurve-package}
\alias{SuperCurve-package}
\docType{package}
\title{Reverse phase protein lysate array analysis}
\description{
  A package for analyzing reverse phase protein lysate arrays (RPPA).
}
\details{
  %% This is absolutely unmaintainable, and it is computed
  %% automatically for installed packages and displayed correctly
  %% either by
  %%    help(package="SuperCurve")
  %% or by
  %%    library(help="SuperCurve")
%  \tabular{ll}{
%    Package: \tab SuperCurve\cr
%    Type: \tab Package\cr
%    Version: \tab 1.3.1\cr
%    Date: \tab 2009-01-08\cr
%    License: \tab See LICENSE for license information\cr
%  }

  For a complete list of functions, use \code{library(help="SuperCurve")}.
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="SuperCurve"))}.
}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\keyword{package}

