\name{RPPASetSummary-class}
\alias{RPPASetSummary-class}
\alias{RPPASetSummary}
\alias{write.summary,RPPASetSummary-method}
\alias{is.RPPASetSummary}
\docType{class}
\title{Class "RPPASetSummary"}
\description{
  The RPPASetSummary class is currently undocumented (TODO).
}
\usage{
RPPASetSummary(rppaset)
is.RPPASetSummary(x)
\S4method{write.summary}{RPPASetSummary}(object,
                                         path,
                                         prefix="supercurve",
                                         graphs=TRUE,
                                         tiffdir=NULL)
}
\arguments{
  \item{object}{object of class \code{RPPASetSummary}}
  \item{path}{character string specifying the path from the current
    directory to the directory containing the files to be processed}
  \item{prefix}{character string used as a prefix on files generated by
    the \code{write.summary} method.}
  \item{graphs}{logical scalar specifying if standard graphs should be produced}
  \item{tiffdir}{character string specifying path to the directory containing
    the original \acronym{TIFF} files}
  \item{rppaset}{object of class \code{RPPASet}}
  \item{x}{object of class \code{RPPASetSummary}}
  \item{\dots}{extra arguments for generic or plotting routines}
}
\details{
  TODO.
}
\section{Objects from the Class}{
  Although objects of the class can (in theory) be created by a direct call
  to \link[methods]{new}, the only realistic method is to use the
  \code{RPPASetSummary} generator function.
}
\section{Slots}{
  \describe{
    \item{\code{raw}:}{TODO}
    \item{\code{ss}:}{TODO}
    \item{\code{medpol}:}{TODO}
  }
}
\section{Methods}{
  \describe{
    \item{summary(object, \dots)}{Creates an object of class
      \code{RPPASetSummary}.}
    \item{write.summary(object, path, prefix, graphs, tiffdir, \dots)}{
      Writes three CSV files, containing raw concentrations, \eqn{R^2} statistics,
      and the polished concentration.}
  }
}
\value{
  The \code{RPPASetSummary} generator returns an object of class
  \code{RPPASetSummary}.

  The \code{is.RPPASetSummary} method returns \code{TRUE} if its
  argument is an object of class \code{RPPASetSummary}.
}
\references{KRC}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\seealso{
  \code{\link{RPPASet}}
}
\examples{
  \dontrun{
parentdir <- file.path("C:", "MyData")
txtdir <- file.path(parentdir, "txt")     # quantification files
outdir <- file.path(parentdir, "results") # output files

designparams <- RPPADesignParams(grouping="blockSample",
                                 center=FALSE,
                                 aliasfile="layoutInfo.tsv",
                                 designfile="slidedesign.tsv")
fitparams <- RPPAFitParams(measure="Mean.Net",
                           method="nlrob",
                           model="cobs",
                           ignoreNegative=FALSE,
                           warnLevel=-1,
                           verbose=FALSE)

fitset <- RPPASet(txtdir,
                  designparams,
                  fitparams)
## If you REALLY want to do this manually. It will be invoked
## automatically if you invoke write.summary(fitset) instead...
rppaset.sum <- summary(fitset)
write.summary(rppaset.sum,
              path=outdir)
  }
}
\keyword{models}
\keyword{regression}
\keyword{nonlinear}
\keyword{nonparametric}
\keyword{robust}

