\name{write.summary-method}
\alias{write.summary}
\alias{write.summary-method}
\alias{write.summary,ANY-method}
\docType{methods}
\title{Method "write.summary"}
\description{
  This method could be thought of as a \code{summary} method that writes
  to disk, saving the summary information in an external format.
}
\usage{
write.summary(object,
              \dots)
}
\arguments{
  \item{object}{object of implementing class}
  \item{\dots}{extra arguments for generic or plotting routines}
}
\author{
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\keyword{methods}

